Introduce `FilterStrategy` to filter an event to be invoked in the listener
