+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/event/SocketEvent.kt
