+++ omisego-sdk/src/main/java/co/omisego/omisego/websocket/event/TransactionConsumptionListener.kt
@@ -3,9 +3,10 @@ package co.omisego.omisego.websocket.event
 import co.omisego.omisego.model.APIError
 import co.omisego.omisego.model.transaction.consumption.TransactionConsumption
 import co.omisego.omisego.websocket.listener.SimpleSocketCustomEventListener
+import co.omisego.omisego.websocket.strategy.FilterStrategy
 
-abstract class TransactionConsumptionListener : SimpleSocketCustomEventListener<SocketEvent<*>>(allowedEvents) {
-
+abstract class TransactionConsumptionListener : SimpleSocketCustomEventListener<SocketEvent<*>>() {
+    override var strategy: FilterStrategy = FilterStrategy.Event(allowedEvents)
     final override fun onSpecificEvent(event: SocketEvent<*>) {
         when (event) {
             is TransactionConsumptionFinalizedEvent -> event.socketReceive.dispatch(
@@ -16,7 +17,7 @@ abstract class TransactionConsumptionListener : SimpleSocketCustomEventListener<
     }
 
     abstract fun onTransactionConsumptionFinalizedSuccess(transactionConsumption: TransactionConsumption)
-    abstract fun onTransactionConsumptionFinalizedFail(transactionConsumption: TransactionConsumption, apiError: APIError)
+    abstract fun onTransactionConsumptionFinalizedFail(transactionConsumption: TransactionConsumption?, apiError: APIError)
 
     companion object {
         private val allowedEvents = listOf(
